+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/ECChronosInternals.java
@@ -239,6 +239,18 @@ public class ECChronosInternals implements Closeable
                         timeUnit.toMillis(timeTaken), successful ? "successful" : "not successful");
             }
         }
+
+        @Override
+        public void repairFailedAttempts(TableReference tableReference)
+        {
+            LOG.debug("Table {} repair failed attempts", tableReference);
+        }
+
+        @Override
+        public void resetRepairFailedAttempts(TableReference tableReference)
+        {
+            LOG.debug("Resetting Table {} repair failed attempts", tableReference);
+        }
     }
 
     private static class NoOpTableStorageState implements TableStorageStates