+++ core.osgi/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/osgi/TableRepairMetricsService.java
@@ -94,6 +94,18 @@ public final class TableRepairMetricsService implements TableRepairMetrics, Tabl
         myDelegateTableRepairMetrics.repairTiming(tableReference, timeTaken, timeUnit, successful);
     }
 
+    @Override
+    public void repairFailedAttempts(TableReference tableReference)
+    {
+        myDelegateTableRepairMetrics.repairFailedAttempts(tableReference);
+    }
+
+    @Override
+    public void resetRepairFailedAttempts(TableReference tableReference)
+    {
+        myDelegateTableRepairMetrics.resetRepairFailedAttempts(tableReference);
+    }
+
     @ObjectClassDefinition
     public @interface Configuration
     {